+++ src/Models/LanguageProxy.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the LanguageProxy class.
+ *
+ * @copyright   Copyright (c) 2023 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-04-08
+ *
+ */
+
+namespace Konekt\Address\Models;
+
+use Konekt\Concord\Proxies\ModelProxy;
+
+class LanguageProxy extends ModelProxy
+{
+}